Add files via upload
The original code compares the counts of 1's (countA and countB) and sorts the elements based on the ascending order of the counts. If countA is equal to countB, it then compares the values of a and b to break the tie.

In this specific case, using countA > countB in the return statement inside the lambda function would result in descending order of counts, rather than ascending order. If we want to sort the elements based on the ascending order of counts, we should return countA < countB instead of countA > countB.